Extracts color fill logic into reusable helper function
Refactors color fill determination logic into a dedicated helper function to improve modularity and code reuse. Updates the test file to use the new helper, reducing redundancy and improving maintainability.
